requirements.txt
paho-mqtt==1.5.1
prometheus-client==0.10.1
pyModbusTCP==0.1.10
